feat [front, kalendarz]: ulepszenie struktury plików i zalezności
dodany zakres w kalendarzu (zamiast jedynie dnia startu)
dodane loading state: zamiast 1. pusty kalendarz -> 2. kalendarz z wydarzeniami
będziemy mieć 1. napis ładowanie wydarzeń -> 2. kalendarz z wydarzeniami
